Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/datetime-local-expected.txt


PASS empty value
PASS datetime-local input value set to 2014-01-01T11:11:11.111 without min/max
FAIL datetime-local input value set to 2014-01-01 11:11:11.111 without min/max assert_equals: expected "2014-01-01T11:11:11.111" but got "2014-01-01 11:11:11.111"
FAIL datetime-local input value set to 2014-01-01 11:11 without min/max assert_equals: expected "2014-01-01T11:11" but got "2014-01-01 11:11"
FAIL datetime-local input value set to 2014-01-01 00:00:00.000 without min/max assert_equals: expected "2014-01-01T00:00" but got "2014-01-01 00:00:00.000"
FAIL datetime-local input value set to 2014-01-0 11:11 without min/max assert_equals: expected "" but got "2014-01-0 11:11"
FAIL datetime-local input value set to 2014-01-01 11:1 without min/max assert_equals: expected "" but got "2014-01-01 11:1"
FAIL invalid datetime-local input value 1 assert_equals: expected "" but got "2014-01-01 11:1d1"
FAIL invalid datetime-local input value 2 assert_equals: expected "" but got "2014-01-01H11:11"
FAIL invalid datetime-local input value 3 assert_equals: expected "" but got "2014-01-01 11:11:"
FAIL invalid datetime-local input value 4 assert_equals: expected "" but got "2014-01-01 11-11"
FAIL invalid datetime-local input value 5 assert_equals: expected "" but got "2014-01-01 11:11:123"
FAIL invalid datetime-local input value 6 assert_equals: expected "" but got "2014-01-01 11:11:12.1234"
FAIL Value >= min attribute assert_equals: expected "2014-01-01T11:12" but got "2014-01-01 11:12"
FAIL Value < min attribute assert_equals: expected "2014-01-01T11:11" but got "2014-01-01 11:10"
FAIL Value <= max attribute assert_equals: expected "2014-01-01T11:10" but got "2014-01-01 11:10"
FAIL Value > max attribute assert_equals: expected "2014-01-01T11:11" but got "2014-01-01 11:12"
